Added Makefiles to build dog.
There is a root directory Makefile which will build all in src.
It also supports install and clean. Install copies all binaries to
the bin directory.

In src the targets are dog ext util (also all, clean and install).

In src\ext there are also targets for individual commands
e.g. vr for vr.com (which is also a target).

In src\util there are targets for the individual utils but also
2 letter aliases:
  hd: head.com
  hh: help.com
  wh: which.com

I might make hd and wh into default aliases in dog.