Added the PP command and documentation.
PP is written in DOG, so PP also functions as a DOG language demo.

Syntax: PP [PROMPT]
Parameters:
           - PP alone displays the formatting for the current DOG prompt.
   PROMPT  - The string displayed indicating that DOG is waiting for a command
             You can use these special characters:
      $$ - the $ sign.
      $_ - space
      $b - vertical bar '|'
      $e - The ESC character (ASCII 27 = 1b in hex)
      $l - Left angle '<'
      $g - Right angle '>'
      $n - New line
      $r - Carriage return
      $t - Tabulator
      $p - Current drive and path
      $c - Current time
Effect: Displays the PROMPT or sets the PROMPT.

In order to get PP to work correctly this also fixes arg parsing for
DOG files and adds the -n OPTION for EH to not use the special formatting
characters.